finalized styling for now
